Time: 5 ms (21.09%), Space: 7.6 MB (16.1%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,13 +1,13 @@
 class Solution {
 public:
     int reverse(int x) {
-        long ans = 0;
-        while(x){
-            int digit = x%10;
+        long long ans = 0;
+        while (x) {
+            int digit = x % 10;
             ans = ans * 10 + digit;
-            x = x/10;
+            x = x / 10;
         }
-        if(ans>INT_MAX||ans<INT_MIN) return 0;
-        return (int)ans;
+        if (ans > INT_MAX || ans < INT_MIN) return 0;
+        return (int) ans;
     }
 };
